pyproject.toml
[project]
name = "dynamic-orchestrator-mcp"
version = "0.1.0"
description = "Dynamic AI System Orchestrator - MCP Service for intelligent request routing"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.19.0",
    "asyncio>=3.11",
    "redis>=5.0.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true